Updated to lock in 250Hz sample rate
Some PX4 logs were not working with the automatic sample time calculator.  Verified that a few files did in fact use the 250Hz sample time but had 1 or more errant samples (where dt was .5~10s, instead of .004s).  Instead of patching the auto calculator just fixed the asfreq() parameter to 250Hz so that all files are consistent.